src/main/java/stream/ProfileStore.java
package stream;

import java.util.List;
import java.util.Comparator;
import java.util.stream.Collectors;

public class ProfileStore {
    public List<Address> collect(List<Profile> profiles) {
        return profiles.stream().map(
                Profile::getAddress
        ).collect(Collectors.toList());
    }

    public List<Address> collectDistinct(List<Profile> profiles) {
        return profiles.stream()
                .map(Profile::getAddress)
                .sorted(Comparator.comparing(Address::getCity))
                .distinct()
                .collect(Collectors.toList());
    }
}
